make http stress workload more consistent over different iterations
use 1 simple pattern instead of a mix of various patterns. we simply
don't do enough requests to have that become a consistent pattern.

Also, we have to filter for which metrics to use.
Without filtering we would get all metrics back, e.g.:

1 some.id.of.a.metric.1
1 some.id.of.a.metric.10
1 some.id.of.a.metric.100
1 some.id.of.a.metric.1000
1 some.id.of.a.metric.10000
1 some.id.of.a.metric.100000
...
1 some.id.of.a.metric.99999

This meant replacing a single char with a '*' could match many more
series then we meant to. e.g. the top one, replacing 1 with * would
match ALL series, resulting in very unfair benchmark runs.
Now we know that we only ever query for 1 or 10 series.